requirements.txt
absl-py==0.9.0
albumentations==0.4.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imgaug==0.2.6
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.0.3
networkx==2.4
numpy
oauthlib==3.1.0
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
Pillow==8.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.7.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image
scikit-learn==0.22.2.post1
scipy
six==1.14.0
sklearn==0.0
tensorboard==2.1.0
torch==1.4.0
torchvision==0.2.2.post3
tqdm==4.27.0
urllib3==1.25.8
Werkzeug==1.0.0
streamlit==0.58